added confirmation dialog when delete record

+++ webui/src/app/components/drives/drive-checklist/drive-checklist.component.ts
@@ -1,10 +1,11 @@
 import { Component, OnInit, ViewChild, ElementRef } from '@angular/core';
-import { MatTableDataSource, MatPaginator, MatSort } from '@angular/material';
+import { MatTableDataSource, MatPaginator, MatSort, MatDialogRef, MatDialog } from '@angular/material';
 import { DriveChecklistModel } from 'src/app/models/drive.model';
 import { Ng4LoadingSpinnerService } from 'ng4-loading-spinner';
 import { CommonService } from 'src/app/common/common.service';
 import { DrivesService } from 'src/app/services/drives.service';
 import { Router, ActivatedRoute } from '@angular/router';
+import { FuseConfirmDialogComponent } from '../../confirm-dialog/confirm-dialog.component';
 
 @Component({
   selector: 'app-drive-checklist',
@@ -17,7 +18,7 @@ export class DriveChecklistComponent implements OnInit {
   addPermission: boolean = true;
   deletePermission: boolean = true;
   userdata: any = JSON.parse(localStorage.getItem('userData'));
-
+  confirmDialogRef: MatDialogRef<FuseConfirmDialogComponent>;
   displayedColumns = ['sno', 'drive', 'measureActivityList', 'displayOrder', 'lowerLimit', 'upperLimit', 'active', 'actions'];
   dataSource: MatTableDataSource<DriveChecklistModel>;
 
@@ -26,7 +27,7 @@ export class DriveChecklistComponent implements OnInit {
   @ViewChild(MatSort, { static: true }) sort: MatSort;
   @ViewChild('filter', { static: true }) filter: ElementRef;
 
-  drivesCheckList:any;
+  drivesCheckList: any;
 
   constructor(
     private spinnerService: Ng4LoadingSpinnerService,
@@ -34,6 +35,7 @@ export class DriveChecklistComponent implements OnInit {
     private drivesService: DrivesService,
     private router: Router,
     private route: ActivatedRoute,
+    public dialog: MatDialog,
   ) { }
 
   ngOnInit() {
@@ -63,26 +65,35 @@ export class DriveChecklistComponent implements OnInit {
 
       this.dataSource = new MatTableDataSource(drive);
       this.dataSource.paginator = this.paginator;
-      this.dataSource.sort = this.sort;            
+      this.dataSource.sort = this.sort;
       this.spinnerService.hide();
     }, error => {
       this.spinnerService.hide();
     });
   }
-  processEditAction(id){
-    this.router.navigate([id],{relativeTo: this.route});
+  processEditAction(id) {
+    this.router.navigate([id], { relativeTo: this.route });
   }
-  delete(id){
-    this.spinnerService.show();
-    this.drivesService.deleteDriveData(id).subscribe(data => {
-      this.spinnerService.hide();
-      this.commonService.showAlertMessage("Deleted Drive Successfully");
-      this.getDrivesData();
-    }, error => {
-      console.log('ERROR >>>');
-      this.spinnerService.hide();
-      this.commonService.showAlertMessage("Drive Deletion Failed.");
-    })
+  delete(id) {
+    this.confirmDialogRef = this.dialog.open(FuseConfirmDialogComponent, {
+      disableClose: false
+    });
+    this.confirmDialogRef.componentInstance.confirmMessage = 'Are you sure you want to delete?';
+    this.confirmDialogRef.afterClosed().subscribe(result => {
+      if (result) {
+        this.spinnerService.show();
+        this.drivesService.deleteDriveData(id).subscribe(data => {
+          this.spinnerService.hide();
+          this.commonService.showAlertMessage("Deleted Drive Successfully");
+          this.getDrivesData();
+        }, error => {
+          console.log('ERROR >>>');
+          this.spinnerService.hide();
+          this.commonService.showAlertMessage("Drive Deletion Failed.");
+        });
+      }
+      this.confirmDialogRef = null;
+    });
   }
 
 }
