+++ webui/src/app/components/drives/drive-target/drive-target.component.ts
@@ -1,10 +1,11 @@
 import { Component, OnInit, ViewChild, ElementRef } from '@angular/core';
 import { DriveTargetModel } from 'src/app/models/drive.model';
-import { MatTableDataSource, MatPaginator, MatSort } from '@angular/material';
+import { MatTableDataSource, MatPaginator, MatSort, MatDialogRef, MatDialog } from '@angular/material';
 import { Ng4LoadingSpinnerService } from 'ng4-loading-spinner';
 import { CommonService } from 'src/app/common/common.service';
 import { DrivesService } from 'src/app/services/drives.service';
 import { Router, ActivatedRoute } from '@angular/router';
+import { FuseConfirmDialogComponent } from '../../confirm-dialog/confirm-dialog.component';
 
 @Component({
   selector: 'app-drive-target',
@@ -16,8 +17,8 @@ export class DriveTargetComponent implements OnInit {
   addPermission: boolean = true;
   deletePermission: boolean = true;
   userdata: any = JSON.parse(localStorage.getItem('userData'));
-
-  displayedColumns = ['sno', 'unitType', 'unitName', 'target', 'poulation', 'driveId','actions'];
+  confirmDialogRef: MatDialogRef<FuseConfirmDialogComponent>;
+  displayedColumns = ['sno', 'unitType', 'unitName', 'target', 'poulation', 'driveId', 'actions'];
   dataSource: MatTableDataSource<DriveTargetModel>;
 
 
@@ -25,7 +26,7 @@ export class DriveTargetComponent implements OnInit {
   @ViewChild(MatSort, { static: true }) sort: MatSort;
   @ViewChild('filter', { static: true }) filter: ElementRef;
 
-  driveTargetList:any;
+  driveTargetList: any;
 
   constructor(
     private spinnerService: Ng4LoadingSpinnerService,
@@ -33,6 +34,7 @@ export class DriveTargetComponent implements OnInit {
     private drivesService: DrivesService,
     private router: Router,
     private route: ActivatedRoute,
+    public dialog: MatDialog,
   ) { }
 
   ngOnInit() {
@@ -52,37 +54,46 @@ export class DriveTargetComponent implements OnInit {
   getDriveTargetData() {
     const driveTarget: DriveTargetModel[] = [];
     this.drivesService.getDriveTargetData().subscribe((data) => {
-      console.log(JSON.stringify(data))
       this.driveTargetList = data;
-       for (let i = 0; i < this.driveTargetList.length; i++) {
+      for (let i = 0; i < this.driveTargetList.length; i++) {
         this.driveTargetList[i].sno = i + 1;
         this.driveTargetList[i].driveId = this.driveTargetList[i].driveId['name'];
         this.driveTargetList[i].depoType = !!this.driveTargetList[i].depotType ? this.driveTargetList[i].depotType['unitType'] : '';
         driveTarget.push(this.driveTargetList[i]);
-      } 
+      }
 
       this.dataSource = new MatTableDataSource(driveTarget);
       this.dataSource.paginator = this.paginator;
-      this.dataSource.sort = this.sort;            
+      this.dataSource.sort = this.sort;
       this.spinnerService.hide();
     }, error => {
       this.spinnerService.hide();
     });
   }
-  processEditAction(id){
-    this.router.navigate([id],{relativeTo: this.route});
+  processEditAction(id) {
+    this.router.navigate([id], { relativeTo: this.route });
   }
-  delete(id){
-    this.spinnerService.show();
-    this.drivesService.deleteDriveTargetData(id).subscribe(data => {
-      console.log(JSON.stringify(data));
-      this.spinnerService.hide();
-      this.commonService.showAlertMessage("Deleted Drive Target Successfully");
-      this.getDriveTargetData();
-    }, error => {
-      console.log('ERROR >>>');
-      this.spinnerService.hide();
-      this.commonService.showAlertMessage("Drive Target Deletion Failed.");
-    })
+  delete(id) {
+    this.confirmDialogRef = this.dialog.open(FuseConfirmDialogComponent, {
+      disableClose: false
+    });
+    this.confirmDialogRef.componentInstance.confirmMessage = 'Are you sure you want to delete?';
+    this.confirmDialogRef.afterClosed().subscribe(result => {
+      if (result) {
+        this.spinnerService.show();
+        this.drivesService.deleteDriveTargetData(id).subscribe(data => {
+          console.log(JSON.stringify(data));
+          this.spinnerService.hide();
+          this.commonService.showAlertMessage("Deleted Drive Target Successfully");
+          this.getDriveTargetData();
+        }, error => {
+          console.log('ERROR >>>');
+          this.spinnerService.hide();
+          this.commonService.showAlertMessage("Drive Target Deletion Failed.");
+        })
+      }
+      this.confirmDialogRef = null;
+    });
   }
+
 }
